added LoggerService
This suggestion from the AI caused an insane amount of work, its initial implementation was full of bugs. I wonder if I should rather have used the logger package instead of basically re-implementing a logger framework here.

Anyways, what I like about this class is the possibility, to read the log content back and thus to check in tests if something specific was logged - that is a very nice feature.