Added our own turtle parser
since there is no turtle/rdf parser available for dart which is maintained and supports the web (at least not as of this writing),
I decided to ask cursor help me implement a turtle parser.

The test with the real world document is still failing, so it is not yet finished, but
since we already did many, many iterations and the code is already quite good, I want to commit it,
before I go on and fix handling of the real world document myself, because cursor failed on this
task even after many iterations for this test alone.