fixed auth and sync, successfully synced to the solid pod for the first time.
Next, we will need to cleanup the code, and think about the data we want to store. Should we really simply store everything in one todos file? Or isn't it better to use directory structure and one file per item? But first I guess we are going to rename this project