First steps of cleaner rdf mapping implementation and service locator cleanup
This commit is certainly influenced by Copilot Agent, but I had to steer it
a lot and had to do quite a few refactorings myself. So this commit
like most others I would say that Agent helped me get a better result,
but it certainly was more like a good sidekick than someone that does good work
by him(it?)self. And I am also not 100% sure how much faster I was than I would have been if I did not have Agent - it does not feel like a huge productivity boost to be honest (but using LLMs to clarify questions and to learn about technology certainly is).

we introduced an extensions mechanism for the service locator, which should make it really clean