Generated Doc Comments
Finally a task, that was executed by the agent fully autonomously and better than I could have done it. This should be helpful for users of the library that know as little about RDF as I did when I decided to implement an app that uses RDF for storing the server on a Solid Pod.